Made changes to indentation and declared common functions in common.h
