+++ C/Readme.md
@@ -23,12 +23,16 @@ It will be overwritten on the next call into the function, so be careful!
 
 ## Usage as a command-line tool
 
-You can also build a command-line version to test the encoder. However, note that it uses `stb_image` to load images,
+You can also build a command-line version to test the encoder and decoder. However, note that it uses `stb_image` to load images,
 which is not really security-hardened, so it is **not** recommended to use this version in production on untrusted data!
 Use one of the integrations instead, which use more robust image loading libraries.
 
 Nevertheless, if you want to try it out quickly, simply run:
 
-	$ make
-	$ ./blurhash 4 3 ../Swift/BlurHashTest/pic1.png
+	$ make blurhash_encoder
+	$ ./blurhash_encoder 4 3 ../Swift/BlurHashTest/pic1.png
 	LaJHjmVu8_~po#smR+a~xaoLWCRj
+
+If you want to try out the decoder, simply run:
+	$ make blurhash_decoder
+	$ ./blurhash_decoder "LaJHjmVu8_~po#smR+a~xaoLWCRj" 32 32 decoded_output.png
